refactor: extract shared NavUser sidebar component
Admin and timesheet shells carried near-identical sidebar user
dropdowns; consolidate into src/components/nav-user.tsx with an
optional onOpenSettings prop for the Einstellungen item.
