+++ src/features/timesheet/components/timesheet-shell.tsx
@@ -2,15 +2,12 @@
 
 import { useMemo, useState } from "react";
 import Image from "next/image";
-import { useRouter } from "next/navigation";
 import {
   CalendarDays,
   CalendarRange,
   ChevronRight,
   Clock,
   LifeBuoy,
-  LogOut,
-  MoreVertical,
   Send,
   Settings,
 } from "lucide-react";
@@ -30,19 +27,9 @@ import {
   SidebarRail,
   SidebarSeparator,
   SidebarTrigger,
-  useSidebar,
 } from "@/components/ui/sidebar";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
 import { Toaster } from "@/components/ui/sonner";
-import { authClient } from "@/lib/auth-client";
-import { getInitials } from "@/lib/utils";
+import { NavUser } from "@/components/nav-user";
 import { BottomTabBar, type TabId } from "./bottom-tab-bar";
 import { NewEntrySheet } from "./new-entry-sheet";
 import { TabDay } from "./tab-day";
@@ -91,80 +78,6 @@ function greet(name: string): string {
   if (h < 14) return `Hallo, ${firstName}`;
   if (h < 18) return `Guten Nachmittag, ${firstName}`;
   return `Guten Abend, ${firstName}`;
-}
-
-function NavUser({
-  user,
-  onOpenSettings,
-}: {
-  user: { name: string; email: string };
-  onOpenSettings: () => void;
-}) {
-  const router = useRouter();
-  const { isMobile } = useSidebar();
-  const initials = getInitials(user.name);
-
-  async function handleSignOut() {
-    await authClient.signOut();
-    router.push("/login");
-    router.refresh();
-  }
-
-  return (
-    <SidebarMenu>
-      <SidebarMenuItem>
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <SidebarMenuButton
-              size="lg"
-              tooltip={user.name}
-              className="data-[state=open]:bg-sidebar-accent data-[state=open]:text-sidebar-accent-foreground"
-            >
-              <div className="grid size-8 shrink-0 place-items-center rounded-lg bg-sidebar-accent text-xs font-semibold text-sidebar-accent-foreground">
-                {initials}
-              </div>
-              <div className="grid flex-1 text-left text-sm leading-tight">
-                <span className="truncate font-medium">{user.name}</span>
-                <span className="truncate text-xs text-muted-foreground">
-                  {user.email}
-                </span>
-              </div>
-              <MoreVertical className="ml-auto size-4" />
-            </SidebarMenuButton>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent
-            className="w-(--radix-dropdown-menu-trigger-width) min-w-56 rounded-lg"
-            side={isMobile ? "bottom" : "right"}
-            align="end"
-            sideOffset={4}
-          >
-            <DropdownMenuLabel className="p-0 font-normal">
-              <div className="flex items-center gap-2 px-1 py-1.5 text-left text-sm">
-                <div className="grid size-8 shrink-0 place-items-center rounded-lg bg-sidebar-accent text-xs font-semibold text-sidebar-accent-foreground">
-                  {initials}
-                </div>
-                <div className="grid flex-1 text-left text-sm leading-tight">
-                  <span className="truncate font-medium">{user.name}</span>
-                  <span className="truncate text-xs text-muted-foreground">
-                    {user.email}
-                  </span>
-                </div>
-              </div>
-            </DropdownMenuLabel>
-            <DropdownMenuSeparator />
-            <DropdownMenuItem onSelect={onOpenSettings}>
-              <Settings />
-              Einstellungen
-            </DropdownMenuItem>
-            <DropdownMenuItem onSelect={handleSignOut}>
-              <LogOut />
-              Abmelden
-            </DropdownMenuItem>
-          </DropdownMenuContent>
-        </DropdownMenu>
-      </SidebarMenuItem>
-    </SidebarMenu>
-  );
 }
 
 export function SchoolAssistantApp({
